+++ programoutput/CHANGES.rst
@@ -1,6 +1,7 @@
 0.5 (in development)
 ====================
 
+- :confval:`programoutput_prompt_template` is interpreted as format string now!
 - Require Python 2.6 now
 - Warn on unexpected return codes instead of raising
   :exc:`~subprocess.CalledProcessError`